processing/cloud_function/D001/preprocess/entry-struct-data/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
google-cloud-run==0.10.14
pymongo==4.8.0
pydantic==2.9.2
pandas==2.2.2
python-magic==0.4.27